rand_distr: add module chi_squared
