+++ rand_distr/src/lib.rs
@@ -101,12 +101,13 @@ pub use rand::distributions::{
 pub use self::beta::{Beta, BetaError};
 pub use self::binomial::{Binomial, Error as BinomialError};
 pub use self::cauchy::{Cauchy, Error as CauchyError};
+pub use self::chi_squared::{ChiSquared, ChiSquaredError};
 #[cfg(feature = "alloc")]
 pub use self::dirichlet::{Dirichlet, Error as DirichletError};
 pub use self::exponential::{Error as ExpError, Exp, Exp1};
 pub use self::fisher_f::{FisherF, FisherFError};
 pub use self::frechet::{Error as FrechetError, Frechet};
-pub use self::gamma::{ChiSquared, ChiSquaredError, Error as GammaError, Gamma};
+pub use self::gamma::{Error as GammaError, Gamma};
 pub use self::geometric::{Error as GeoError, Geometric, StandardGeometric};
 pub use self::gumbel::{Error as GumbelError, Gumbel};
 pub use self::hypergeometric::{Error as HyperGeoError, Hypergeometric};
@@ -195,6 +196,7 @@ pub mod weighted_tree;
 mod beta;
 mod binomial;
 mod cauchy;
+mod chi_squared;
 mod dirichlet;
 mod exponential;
 mod fisher_f;
